backend/pyproject.toml
[tool.poetry]
name = "voice-anime-fighter-backend"
version = "0.1.0"
description = "FastAPI backend for Voice-Anime-Fighter game"
authors = ["Developer"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-socketio = "^5.11.0"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
redis = "^5.0.1"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
azure-cognitiveservices-speech = "^1.35.0"
librosa = "^0.10.1"
numpy = "^1.26.3"
scipy = "^1.12.0"
aiofiles = "^23.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
